feat: Catch with prior PR that now uses short instead of integer for small numbers

+++ core/src/test/scala/com/databricks/labs/remorph/parsers/tsql/TSqlFunctionSpec.scala
@@ -257,7 +257,7 @@ class TSqlFunctionSpec extends AnyWordSpec with TSqlParserTestCommon with Matche
       ir.Window(
         ir.CallFunction(
           "LEAD",
-          Seq(simplyNamedColumn("salary"), ir.Literal(integer = Some(1)), ir.Literal(boolean = Some(true)))),
+          Seq(simplyNamedColumn("salary"), ir.Literal(short = Some(1)), ir.Literal(boolean = Some(true)))),
         Seq(simplyNamedColumn("department_id")),
         Seq(ir.SortOrder(simplyNamedColumn("employee_id"), ir.DescendingSortDirection, ir.SortNullsUnspecified)),
         None))
